12/pyproject.toml
[project]
name = "12"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.3.0",
    "fastapi>=0.123.0",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
